Began converting over the data lookup to support the pastLocationsList on the main screen. Making this push as a test to the new Dev branch.
