+++ app/src/main/java/com/imobile3/taylor/imobile3_weather_app/fragments/MainFragment.java
@@ -27,15 +27,26 @@ import android.widget.EditText;
 import android.widget.ListView;
 import android.widget.TextView;
 
+import com.imobile3.taylor.imobile3_weather_app.HttpJSONRequest;
 import com.imobile3.taylor.imobile3_weather_app.LocationLookup;
 import com.imobile3.taylor.imobile3_weather_app.ForecastLocationListener;
 import com.imobile3.taylor.imobile3_weather_app.R;
+import com.imobile3.taylor.imobile3_weather_app.activities.MainActivity;
 import com.imobile3.taylor.imobile3_weather_app.activities.SimpleForecastActivity;
 import com.imobile3.taylor.imobile3_weather_app.adapters.PastLocationsAdapter;
 import com.imobile3.taylor.imobile3_weather_app.interfaces.LocationDataTaskListener;
+import com.imobile3.taylor.imobile3_weather_app.interfaces.WeatherDataTaskListener;
+import com.imobile3.taylor.imobile3_weather_app.models.DetailedWeatherItem;
 import com.imobile3.taylor.imobile3_weather_app.models.LocationData;
+import com.imobile3.taylor.imobile3_weather_app.models.WeatherItem;
+import com.imobile3.taylor.imobile3_weather_app.utilities.Utils;
 
-import java.util.HashSet;
+import org.json.JSONArray;
+import org.json.JSONException;
+import org.json.JSONObject;
+
+import java.io.IOException;
+import java.util.ArrayList;
 
 /**
  * Created by Taylor Parrish on 8/29/2016.
@@ -49,7 +60,7 @@ import java.util.HashSet;
  * Issue 2: If Fragment is recreated after failed location lookup sharedpreferences is erased.
  * Possible Solutions: Need to handle exception better. RevertLocation method probably the issue.
  */
-public class MainFragment extends Fragment implements LocationDataTaskListener {
+public class MainFragment extends Fragment implements LocationDataTaskListener, WeatherDataTaskListener {
     private static final String CLASS_TAG = MainFragment.class.getSimpleName();
     private static final boolean DEBUG = true;
 
@@ -59,6 +70,7 @@ public class MainFragment extends Fragment implements LocationDataTaskListener {
     private LocationManager mLocationManager;
     private ProgressDialog mProgressDialog;
     private SharedPreferences mSharedPreferences;
+    private String mLocation;
 
     @Override
     public void onCreate(Bundle savedInstanceState) {
@@ -127,6 +139,11 @@ public class MainFragment extends Fragment implements LocationDataTaskListener {
     }
 
     @Override
+    public void onTaskFinished(ArrayList<WeatherItem> weatherItems) {
+
+    }
+
+    @Override
     public void onTaskFinished(LocationData locationData) {
         if (DEBUG) Log.d(CLASS_TAG, "onTaskFinished()");
 
@@ -143,15 +160,13 @@ public class MainFragment extends Fragment implements LocationDataTaskListener {
         final String formatted_address = locationData.getFormattedAddress();
         final String city = locationData.getCity();
         final String location = latitude + "," + longitude;
+        mLocation = location;
 
         AlertDialog.Builder builder = new AlertDialog.Builder(getActivity());
 
         builder.setPositiveButton("Submit", new DialogInterface.OnClickListener() {
             public void onClick(DialogInterface dialog, int id) {
-                Intent mSimpleForecastIntent = new Intent(getActivity(), SimpleForecastActivity.class);
-                mSimpleForecastIntent.putExtra(TAG_EXTRA_LOCATION, location);
-                mSharedPreferences.edit().putString(city, location).apply();
-                startActivity(mSimpleForecastIntent);
+               new JSONParser(MainFragment.this).execute();
             }
         });
         builder.setNegativeButton("Cancel", new DialogInterface.OnClickListener() {
@@ -297,6 +312,118 @@ public class MainFragment extends Fragment implements LocationDataTaskListener {
         @Override
         public void onPostExecute(LocationData locationData) {
             listener.onTaskFinished(locationData);
+        }
+    }
+
+    private class JSONParser extends AsyncTask<String, String, JSONObject> {
+        private WeatherDataTaskListener mListener;
+
+        public JSONParser(WeatherDataTaskListener listener) {
+            this.mListener = listener;
+        }
+
+        @Override
+        public void onPreExecute() {
+            super.onPreExecute();
+            mListener.onTaskStarted();
+        }
+
+        @Override
+        public JSONObject doInBackground(String... args) {
+            final String WUNDERGROUND_API_KEY = "20a88f5fc4c597d7";
+
+            //URL for WUnderground API Call
+            String url = "http://api.wunderground.com/api/" + WUNDERGROUND_API_KEY
+                    + "/forecast/q/" + mLocation + ".json";
+
+            try {
+                return new HttpJSONRequest().getJSONFromUrl(url);
+            } catch (IOException | JSONException e) {
+                //Need to handle this properly. Network may not be enabled/working.
+                e.printStackTrace();
+                revertToMainPage();
+            }
+            return null;
+        }
+
+        @Override
+        public void onPostExecute(JSONObject JSONWeatherData) {
+            ArrayList<WeatherItem> weatherItems = parseJSONWeatherData(JSONWeatherData);
+            mListener.onTaskFinished(weatherItems);
+        }
+
+        //Parses JSON Data into its respective model objects
+        private ArrayList<WeatherItem> parseJSONWeatherData(JSONObject JSONWeatherData) {
+            ArrayList<WeatherItem> weatherItems = new ArrayList<>();
+            ArrayList<DetailedWeatherItem> detailWeatherItems = new ArrayList<>();
+
+            try {
+                // Getting JSON Wunderground Simpleforecast data
+                JSONArray simpleforecastData = JSONWeatherData.getJSONObject("forecast")
+                        .getJSONObject("simpleforecast").getJSONArray("forecastday");
+                // Getting JSON Wunderground detail forecast data
+                JSONArray detailforecastData = JSONWeatherData.getJSONObject("forecast")
+                        .getJSONObject("txt_forecast").getJSONArray("forecastday");
+
+                //Parse the corresponding data into WeatherItem object model
+                parseSimpleForecastDataModel(weatherItems, simpleforecastData);
+                //Parse the corresponding data into WeatherItem object model
+                parseDetailForecastDataModel(detailWeatherItems, detailforecastData);
+                //Set List of DetailWeatherItems to corresponding WeatherItem model
+                setWeatherItemDetails(weatherItems, detailWeatherItems);
+            } catch (JSONException e) {
+                e.printStackTrace();
+                revertToMainPage();
+            }
+            return weatherItems;
+        }
+
+        private void setWeatherItemDetails(ArrayList<WeatherItem> weatherItems,
+                                           ArrayList<DetailedWeatherItem> detailWeatherItems) {
+            //Find a more efficient way to do this?
+            for (int i = 0; i < weatherItems.size(); i++) {
+                for (int j = 0; j < detailWeatherItems.size(); j++) {
+                    String weekday = Utils.getFirstWord(detailWeatherItems.get(j).getWeekday());
+                    if (weatherItems.get(i).getWeekday().equals(weekday)) {
+                        weatherItems.get(i).addDetailWeatherItem(detailWeatherItems.get(j));
+                    }
+                }
+            }
+        }
+
+        private void parseDetailForecastDataModel(ArrayList<DetailedWeatherItem> detailWeatherItems,
+                                                  JSONArray detailforecastData) throws JSONException {
+            JSONObject detailData;
+            for (int i = 0; i < detailforecastData.length(); i++) {
+                detailData = detailforecastData.getJSONObject(i);
+
+                String weekday = detailData.getString("title");
+                String description = detailData.getString("fcttext");
+                String pop = detailData.getString("pop");
+
+                detailWeatherItems.add(new DetailedWeatherItem(weekday, description, pop));
+            }
+        }
+
+        private void parseSimpleForecastDataModel(ArrayList<WeatherItem> weatherItems,
+                                                  JSONArray simpleforecastData) throws JSONException {
+            JSONObject simpleData;
+            for (int i = 0; i < simpleforecastData.length(); i++) {
+                simpleData = simpleforecastData.getJSONObject(i);
+
+                String weekday = simpleData.getJSONObject("date").getString("weekday");
+                String conditions = simpleData.getString("conditions");
+                String high = simpleData.getJSONObject("high").getString("fahrenheit") + "˚ F";
+                String low = simpleData.getJSONObject("low").getString("fahrenheit") + "˚ F";
+
+                weatherItems.add(new WeatherItem(weekday, conditions, high, low));
+            }
+        }
+
+        private void revertToMainPage() {
+            Utils.showToast(getActivity(),
+                    "Something went wrong, try another location. Make sure you have internet connection.");
+            startActivity(new Intent(getActivity(), MainActivity.class));
         }
     }
 }
